+++ CHANGELOG.md
@@ -24,9 +24,10 @@
 - Key definitions
   - [BREAKING CHANGE] Switch launcher key from config `"bract.core.launcher"` to context `:bract.core/launcher`
   - Change launcher parser from `str->var` to `any->fn`
+  - [BREAKING CHANGE] Remove exports config keydef (use variable instead)
+- Inducers
+  - [BREAKING CHANGE] Have `export-as-sysprops` and `unexport-sysprops` accept export-key/value arguments
 - [Todo - BREAKING CHANGE] Drop bract.core.type and bract.core.impl in favour of a multimethod (for flexibility)
-- [Todo - BREAKING CHANGE] Switch launcher from config to context value
-- [Todo - BREAKING CHANGE] Remove exports config keydef (use variable instead)
 - [Todo] Add `bract.core.main` namespace for non-CLI apps
 - [Todo] Introduce inducer-list key definition for
   - pre-config